// src/lib/prisma.d.ts
declare const prisma: import("../generated/prisma/internal/class.js").PrismaClient<"error" | "warn", import("../generated/prisma/internal/prismaNamespace.js").GlobalOmitConfig | undefined, import("@prisma/client/runtime/client").DefaultArgs>;
export default prisma;
//# sourceMappingURL=prisma.d.ts.map